Workshop/requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-ai-ml==1.13.0
azure-ai-vision==0.15.1b1
azure-identity==1.12.0
azure-search-documents==11.4.0b9
azure-storage-blob==12.16.0
aiohttp==3.8.6
beautifulsoup4==4.12.2
blobfile==2.0.1
bs4==0.0.1
#fastapi==0.104.1
Faker==19.1.0
joblib==1.2.0
langchain==0.0.347
langchain-core==0.0.11
llama-index==0.9.8.post1
openai==1.12.0
pandas==1.5.3
pathos==0.2.9
pdfminer.six==20221105
Pillow==10.1.0
promptflow[azure]==1.1.0
promptflow-tools==1.0.0
regex==2023.6.3
requests==2.31.0
ratelimit==2.2.1
requests-oauthlib==1.3.1
SQLAlchemy==2.0.23
starlette==0.30.0
semantic-kernel==0.4.1.dev0
sec-edgar-downloader==5.0.2
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
urllib3==2.0.4
unstructured==0.8.1
yfinance==0.2.30
wrapt==1.14.1
llama_hub==0.0.50
wandb==0.16.0
poppler-utils==0.1.0
pymupdf==1.23.26
bcolors==1.0.4
